- Moved defaultWebDAVAccount to osaf.sharing, and out of //parcels so the dump code will pick it up - Added more sharing-related dump/reload translation - SmartCollections without colors now dump/reload - TestDumpReload now tests against more kinds
git-svn-id: http://svn.osafoundation.org/chandler/trunk@13794 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/application/dialogs/AccountPreferences.py
@@ -817,7 +817,8 @@ class AccountPreferencesDialog(wx.Dialog):
             if account['isNew']:
                 uuid = account['item']
                 item = self.rv.findUUID(uuid)
-                item.password.delete()
+                if hasattr(item, 'password'):
+                    item.password.delete()
                 item.delete(recursive=True)
 
             elif account['type'] == "INCOMING":
